Deployment/requirements.txt
streamlit
pandas
plotly
seaborn
matplotlib
Pillow
scikit-learn
imbalanced-learn
